requirements.txt
django
psycopg2
djangorestframework
django-cors-headers
djangorestframework-simplejwt